refactor(auth): extract BackupCredentials shared component
SaveAuthPrompt's backup view and BackupModal rendered near-identical
credential blocks (~80 lines each). Extract BackupCredentials.svelte
with an idPrefix prop and use it from both. Also fixes the missing
unavailable-password fallback in SaveAuthPrompt.
